pyproject.toml
[project]
name = "book"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "flask-cors>=6.0.0",
    "pymongo>=4.13.0",
    "python-dotenv>=1.0.1",
]
